Refactor code to have a proper package structure
