+++ countryguessr/CLIP_Embeddings/Image/.gitignore
