+++ countryguessr/CLIP_Embeddings/Prompt/.gitignore
