+++ countryguessr/CLIP_Experiment/clip_results/.gitignore
