+++ countryguessr/CLIP_Experiment/result_accuracy/.gitignore
