+++ countryguessr/__init__.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+"""
+countryguessr.__init__
+----------------------
+
+Top level of package containing code for the
+Good Pracices in Machine Learning Research project.
+"""
+# Imports
+# Built-in
+
+# Local
+from __version__ import __version__  # noqa: F401
+
+# 3r-party
