+++ countryguessr/__version__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""
+countryguessr.__version__
+-------------------------
+
+Version Tracker.
+
+:copyright: Cognitive Systems Lab, 2025
+"""
+
+VERSION = (0, 0, 1)
+
+__version__ = ".".join(map(str, VERSION))
