+++ countryguessr/data_collection/aerial/aerial_data_collection.py
