+++ countryguessr/data_collection/aerial/interior_boxes.csv
