+++ countryguessr/data_collection/aerial/island_bounding_boxes.csv
