+++ countryguessr/data_collection/aerial/ne_110m_admin_0_countries.json
