+++ countryguessr/data_collection/tourist/tourist_data_collection.py
