+++ countryguessr/data_collection/tourist/url_paths_for_tourist_collection.json
