+++ countryguessr/finetuning/create_datasets_from_embeddings.py
