+++ countryguessr/finetuning/finetuning/runs/seed_4808/geo_strongly_balanced/starting_regional_loss_portion-0.0/regional_loss_decline-1.0/20240523_161648/events.out.tfevents.1716473808.rechenknecht8.379085.0
