+++ countryguessr/utils/analyzation_tools.py
