+++ countryguessr/utils/pdf.py
